server/requirements.txt
# This file is used with Docker development environments. It is not used if you are using conda virtual environments for development.

aiofiles==0.5.0
cryptography==42.0.5
fastapi==0.110.0
fsspec==2021.5.0
geopandas==0.9.0
pandas==2.2.2
# plotly==4.14.3
psutil==5.8.0
psycopg2-binary==2.9.1
pydantic==2.6.4
pydantic_core==2.16.3
pyjwt==2.0.1
PyMySQL==1.0.3
python-dotenv==1.0.1
python-multipart==0.0.5
s3fs==2021.5.0
scipy==1.13.1
SQLAlchemy==2.0.28
uvicorn==0.29.0
